requirements.txt
Django==5.2.5
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1

psycopg2-binary==2.9.6
redis==4.5.4
django-redis==5.2.0

drf-yasg==1.21.5
django-filter==24.2.0
django-cors-headers==4.4.0

python-decouple==3.8
PyJWT==2.10.1
whitenoise==6.4.0

gunicorn==20.1.0

asgiref==3.9.1
sqlparse==0.5.3